Fix: order termination of ust_thread_manage_notify
The following scenario lead to a deadlock:
 - An application is being registered sessiond side (thread_dispatch_ust_registration).
 - ust_thread_manage_notify is already terminated via quit pipe, hence
   not listening on the notify socket.
 - The application is trying to register an event (lttng_event_create->ustcomm_register_event) via the notify socket.
 - The notify socket is never closed.

The application hang on the communication via the notify socket since
no one is responding, sessiond is waiting for a reply on the
ustctl_register_done call.

Forcing an order of termination on ust_thread_manage_notify and
thread_dispatch_ust_registration guarantees that the socket will be
closed and allow for regular execution. This guarantee is provided by
the fact that no further registration can happen in the absence of the
registration dispatch thread and all current applications will be present
when the ust_thread_manage_notify perform its cleanup.

This can be reproduced using the sessiond_teardown_active_session
scenario provided by [1].

[1] https://github.com/PSRCode/lttng-stress

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>