HACK 101: this is working...
This is at least one way to do it...

Honestly this is most probably enought and ensure that all client version
is supported (2.8/2.9/2.10/2.11). I think, will add testing to lttng-ivc
to test it.

Great care is done do ensuere that we do no it a corner case on the
parsing.

In a better world, we would receive the hostname, datetime and session name
in separate chunk and "glue" them at the leaf level (either
lttng-sessiond or consumerd if local). This touch a shitton of places
and still require "parsing" of session name in case of automatic naming
"auto"... and also require weird handling on non-live streaming
session since the information is a hack on the lttng-ctl side.

Will at least do the complete testing based on this functionnal stuff.

If we plan on changing stuff and break thing we will at least have a
working start.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>