Move to regex handling of second token to take care of custom url path.
One can set a "extra path" during at the set-url and other remote
mechanism. This also handle the add-output snapshot scenario.

Note that presently when adding a snapshot output via add-output, the
session name is simply not present on the relayd filesystem. While when
putting the session in snapshot mode at the create time (--snapshot) the
session name is set in the snapshot output.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>