FOR TESTING PURPOSE ONLY Holding the stream lock does not equate to having data pending
The live timer can hold the stream lock while sending empty beacon. An empty beacon
does not mean that data is still pending for the stream.

A more evolved way to do this is to indroduce a state indicating if we are
doing an operation related to data handling. Looking at the state on a failed
attempt to lock could help when sending huge amount of data, since we would be
able to return early.