FOR TESTING PURPOSE ONLY: Bound maximum data read to RECV_DATA_BUFFER_SIZE per iteration
Do not consume everything all at once even if there is data left on the
socket. This is to provide fairness to the overall data handling of all
connections.

It also provide an bounded processing time of data processing iteration.

This might not be good for perfomance. More benchmark should be done
regarding this.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>