Implement lttng_consumer_clear_channel
This function is responsible for performing all operation needed to
perform a clear action on a given channel.

It only support performing clear operation on unmonitored channel
(snapshot mode) for now.

Do do so we flush and clear all the channel streams.

We use an active flush (consumer_flush_buffer(..., 1) since we consider
the producer active at all time. No reason so far to check for the
quiescent state of the channel. Might need to be revisited.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>