Handle registration of event-rule trigger
Add a ht to the state to keep track of all event-rule based trigger.

We also add a "token" generator which act as a unique id for event-rule
based trigger and to be used as a key in the hash table. For now this is a
simple uint64 incremented counter. This counter does not need
synchronization afaik due to the nature of the thread command handling.

TODO: merge the key matching fix for uint64_t key later in this patchset

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Iee42539f0a664ead5ca03534549c6bbd5e505953