WIP: base struct for event_rule condition evaluation
For now we simply "pass" the trigger name that leaded to the notification
in the case of an event rule condition.

Later this would be replaced with a possible captured payload if any.

This is simply a placeholder for now until we decide what exactly is
pertinent for the notification client to receive in the case of even rule
hit based commmunication.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I56de89aab397704f0754fac7dd31fd67f0739ff0