tests: append to AM_CFLAGS instead of overriding it
The Makefiles modified by this patch currently override the AM_CFLAGS
value, which means that anything put in AM_CFLAGS by configure (for
example, the warning flags) is lost.

I believe the intention is to add some flags to CFLAGS, so modify them
such that they append instead of override.

notification/Makefile.am overrides AM_LDFLAGS with nothing.  It feels
like it was not the intention to actually clear that variable, but
rather that it is just the by-product of a copy paste.  If it was really
the intention to clear the value of AM_LDFLAGS, there would have been a
comment to explain it, right?

Change-Id: I0c90e259ef652762d3e36b486e56513e3bbfd3a2
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>