common: rename filter bytecode types
The filter bytecode is going to be used for implementing capture
expressions, so it's not going to be specific to filter.  This patch
renames the types related to bytecode to remove "filter" from their
names.

All the changes start in bytecode/bytecode.h, the changes in the other
files are only fallouts.

- `enum filter_op` is renamed `enum bytecode_op`.
- `FILTER_OP_*` enumerators are renamed `BYTECODE_OP_*`.
- The `filter_opcode_t` typedef is renamed `bytecode_opcode_t`.
- `struct lttng_filter_bytecode_alloc` is renamed `struct lttng_bytecode_alloc`.
- `struct lttng_filter_bytecode` is renamed `struct lttng_bytecode`.

Change-Id: Iefdc667f3fbf715f868926c90108c6a5580b6f25
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Depends-on: lttng-ust: I5a800fc92e588c2a6a0e26282b0ad5f31c044479