condition: lttng_capture_descriptor struct
The capture_index corresponds to the index at wich the capture is
expected on reception of the trigger payload from the tracer.

This field will be populated on reception of the trigger on the sessiond
side and is internal use only.

This is necessary due to the fact that a single condition can have
multiple capture actions associated to it with capture descriptor duplicate.
We plan on forming an ordered set of these capture descriptors and
configure only the set for the enabler. This ensure that we capture only
what is necessary on the tracer side. On reception, each capture
descriptor will know the index at which the payload field it corresponds
to is situated.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Ie3309511347818f88c61c750723f21c73d82f80b