sessiond: Add msgpack-c 3.3.0 to the tree
Add a copy of the msgpack-c v3.3.0 release to the tree. Only the files
of the 'C' version of the library are included and the only
modifications are to add the 'vendor/' prefix to the includes to make
sure we don't interact with a system installed version of msgpack and
the addition of lttng-config.h to glue in our endianess compat code.

Currently, the project is built against the system's libmsgpack-c.
However, this isn't viable for release since the library has had
breaking changes (sometimes without bumping the major version) every
year since its initial release.

Given the variety of distribution versions we have to support, I don't
feel comfortable commiting to support all possible versions of this
library.

Nonetheless, this appears to be the most robust C msgpack implementation
and it is permissively licensed, so it is simply easier to add a copy to
the tree.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Change-Id: Ic3d3fedff9a24ca81ba66cc0c970ba0f8f0cba9a