Fix: test: base-path tests are never run
Observed issue
==============

The base-path tests are never run during `make check`.

Cause
=====

The test file is not present in the test list.

Solution
========

Add the test file to the test list

Known drawbacks
=========

One of the substest is currently failing.

  not ok 20 - Snapshot recorded
  #   Failed test 'Snapshot recorded'
  #   in .//../../../utils/utils.sh:lttng_snapshot_record() at line 1468.
  ok 21 - Destroy session ust_app_snapshot_base_path
  not ok 22 - Validate trace for event tp:tptest

References
==========

2a1668643ca94195d5c3889d0337e19165805a42 [1] introduces the tests.

The test file was never part of the test list.

[1] https://github.com/lttng/lttng-tools/commit/2a1668643ca94195d5c3889d0337e19165805a42

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I9a8040141c29e391363b17954a20155abdb3cc04