Fix: registry map key mapping
Observed issue
==============

For the following scenario:

  lttng create test
  lttng add-map -u test
  lttng start
  lttng add-trigger --condition on-event --action incr-value --session test --map  --key my_key
  ./sample
  ./sample

On the second app the sessiond assert on lttng_ht_add_unique_u64 in
ust_registry_map_add_token_key_mapping

Cause
=====

The trigger tracer token does not change across apps so the tracer token
to key mapping might exists at that point.

Solution
========

Check if the mapping already exists and validate that the associated key
is the same.

TODO: discussion: do we want to move that check in the caller and only
perform that check for buffer UID considering that in PID each apps have
its own maps that have its own mapping?

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Ic407ed981ba10830d4fe593cb823bf2bb0dff8dc