Fix: use tracer_token for trace_ust_find_event
To support multiple trigger with the same condition but with different
incr-value under the same session and same map but with a different key
setting. This ensure that we discriminate on the tracer_token derived
from the lttng trigger object.

We pass 0 for regular *to buffer* event. This ensure that it is not a
discriminant for now.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Ie46373496b9f5ba8ba648d8170717e9c6e810d21