Use the lttng_map_key for ust enabler comparison and fix up the lifetime of lttng_map_key
Since that ust enabler cannot be removed we need to be able to re-enable them if
they match across add/remove trigger. To do this we use the
lttng_map_key as a supplementary discriminant when looking for them in
the map.

Also this mean that we have to keep around the lttng_map_key object
separately from the trigger/actions that initially provided the
lttng-map-key. We use the lttng_map_key refcounting for this.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: If15036cd65ac0c3b7553b9e528bc550567edce94