Fix: data reception is called event when pnly FDS are expected for reception
In non blocking mode, the FDs reception code can hit EAGAIN. In such
cases, skip the data reception completely when performing, again, the
reception.

The same behavior could have been implemented with a noop at the
`lttcomm_recv_unix_sock_non_block` level when the passed `len` is equal
to 0. The handling of it at the notification thread reception level was
chosen to be explicit in regards to the expected behaviour and not rely
on implicit functionality of `lttcomm_recv_unix_sock_non_block`.

While there, augment assertion for code in unix.c

Change-Id: I0fa6df1c9630caa6e877a39372bdfd3d3a01b648
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>